refactor: rename module core to files
