+++ src/commands/install.rs
@@ -4,8 +4,8 @@ use std::path::{Path, PathBuf};
 
 use crate::{
     commands::{self, download::download_binary},
-    core::{archives, filesys},
     datadirs,
+    files::{archives, filesys},
     github::client::{get_asset, get_release},
     selector::is_env_compatible,
     semver_utils::SemverStringPrefix,
