+++ src/files/magic.rs
